+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### Added
+- Add support for TLS
+
+### Changed
+- Update Kotlin, Alexa Skill SDK and Jetty
+
 ## [1.0] - 2017-02-11
 
 ### Added